+++ spec/integration/resource/api600/synergy/sas_logical_interconnect/create_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::Synergy::SASLogicalInterconnect
+klass = OneviewSDK::API600::Synergy::SASLogicalInterconnect
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  let(:current_client) { $client_500_synergy }
-  include_examples 'SASLogicalInterconnectCreateExample', 'integration api500 context'
+  let(:current_client) { $client_600_synergy }
+  include_examples 'SASLogicalInterconnectCreateExample', 'integration api600 context'
 end